Eliminé componente AgendarClases viejo, cambié de nombre al componente Clases como AgendarClases y cambié el estilo de AgendarClases nuevo
