+++ gestor-gimnasio-app/src/components/agendar-clases/AgendarClases.css
@@ -1,80 +1,83 @@
-.agendar-container {
-  max-width: 800px;
-  margin: 40px auto;
-  padding: 0 16px;
+.clases-table .MuiTableCell-root {
+  text-align: center;
 }
 
-.agendar-title {
-  font-size: 2rem;
-  font-weight: bold;
-  margin-bottom: 28px;
-}
-
-.agendar-lista {
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-}
-
-.agendar-clase {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
+/* Contenedor de la tabla */
+.clases-table {
   background: #fff;
-  border-radius: 5px;
-  box-shadow: 1px 2px 6px rgba(0,0,0,0.13);
-  border: 1px solid #e0e0e0;
-  padding: 18px 24px;
-  transition: box-shadow 0.2s;
+  border-radius: 12px;
+  box-shadow: 0 8px 32px rgba(60, 60, 60, 0.22);
+  margin: 32px auto;
+  max-width: 900px;
 }
 
-.agendar-clase.completo {
-  background: #bdbdbd;
+/* Cabecera de la tabla */
+.cabecera-tabla {
+  background: #f5f5f5;
+}
+
+.cabecera-tabla .MuiTableCell-head {
   color: #222;
-}
-
-.agendar-info {
-  flex: 1;
-}
-
-.agendar-fecha {
+  font-weight: 600;
   font-size: 1.1rem;
-  margin-bottom: 4px;
+  border-bottom: 2px solid #e0e0e0;
+  text-align: center;
 }
 
-.agendar-detalle {
+/* Celdas de la tabla */
+.MuiTableCell-root {
+  text-align: center;
   font-size: 1rem;
-  color: #444;
+  border-bottom: 1px solid #e0e0e0;
 }
 
-.agendar-clase.completo .agendar-detalle {
-  color: #444;
+/* Fila al pasar el mouse */
+.MuiTableRow-root:hover {
+  background: #f8fafc;
 }
 
-.agendar-accion {
-  margin-left: 24px;
-  display: flex;
-  align-items: center;
-}
-
-.agendar-btn {
-  background: #000;
-  color: #fff;
-  border: none;
-  border-radius: 7px;
-  padding: 10px 32px;
-  font-size: 1rem;
-  font-weight: bold;
-  cursor: pointer;
+/* Botón negro estilo login */
+.boton-negro {
+  background: #000 !important;
+  color: #fff !important;
+  border-radius: 8px !important;
+  font-weight: 500;
+  box-shadow: 0 2px 6px rgba(60, 60, 60, 0.08);
+  padding: 10px 0;
+  width: 170px;
   transition: background 0.2s;
+  border: none !important;
 }
 
-.agendar-btn:hover {
-  background: #222;
+.boton-negro:hover,
+.boton-negro:focus {
+  background: #363636 !important;
+  color: #fff !important;
 }
 
-.agendar-completo {
-  font-size: 1.2rem;
-  font-weight: bold;
-  letter-spacing: 1px;
+/* Botón outlined negro */
+.boton-outlined-negro {
+  border: 2px solid #000 !important;
+  color: #000 !important;
+  background: #fff !important;
+  border-radius: 8px !important;
+  font-weight: 500;
+  padding: 10px 0;
+  width: 170px;
+  transition: background 0.2s, border-color 0.2s;
+}
+
+.boton-outlined-negro:hover,
+.boton-outlined-negro:focus {
+  border-color: #222222 !important;
+  background: #f5f5f5 !important;
+  color: #000 !important;
+}
+
+.titulo-clases {
+  font-size: 2rem;
+  font-weight: 700;
+  margin: 24px 0 12px 0;
+  font-family: "Roboto", Arial, sans-serif;
+  letter-spacing: -1px;
 }
