--- gestor-gimnasio-app/src/components/clases/Clases.css
@@ -1,3 +1,0 @@
-.clases-table .MuiTableCell-root {
-    text-align: center;
-}
